requirements.txt
torch>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
flask>=3.0.0
flask-socketio>=5.3.0
